pyproject.toml
[tool.poetry]
name = "lm-council"
version = "0.1.0"
description = "A framework for LLM evaluation through multi-judge councils"
authors = ["Justin Zhao <justinxzhao@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/justinzhao/llm-council-public"
repository = "https://github.com/justinzhao/llm-council-public"
keywords = ["llm", "evaluation", "judging", "council", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = ">=3.8"
hypothesis-jsonschema = "^0.23.1"
openai = "^1.63.0"
datasets = ">=2.14.0"
choix = "^0.3.6"
seaborn = "^0.13.2"
scikit-learn = "*"
jsonlines = "^4.0.0"
tiktoken = "*"
python-dotenv = "^1.0.1"
wordfreq = "^3.1.1"
nltk = "^3.9.1"
ete3 = {extras = ["visualization"], version = "^3.1.3"}
pyqt5 = "^5.15.11"
aiolimiter = "^1.2.1"
pydantic = "^2.0"
streamlit = "^1.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
